requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
frozenlist==1.5.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
Js2Py==0.74
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
openai==0.28.0
packaging==24.2
pipwin==0.5.2
propcache==0.3.0
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyjsparser==2.7.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
starlette==0.46.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
yarl==1.18.3
